apps/docx-cli/pyproject.toml
[project]
name = "docx-cli"
version = "0.1.0"
description = "CLI for editing DOCX files via JSON-based update commands"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9.0",
  "python-docx>=1.1.0",
  "lxml>=5.0.0",
  "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "ruff>=0.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docx_cli"]

[tool.ruff]
line-length = 100
target-version = "py310"
